Break out claim organisation invite code into its own class.
