+++ Frontend/CO.CDP.OrganisationApp/Pages/ClaimOrganisationInvite.cshtml
@@ -1,0 +1,2 @@
+@page "/claim-organisation-invite/{personInviteId}"
+@model ClaimOrganisationInviteModel
